+++ ChapeauxUI/DeleteMenuItem.cs
@@ -33,7 +33,7 @@ namespace ChapeauxUI
                 Item itemToDelete = new Item()
                 {
                     Name = txtName.Text,
-                    Price = double.Parse(txtPrice.Text),
+                    Price = decimal.Parse(txtPrice.Text),
                     Stock = int.Parse(txtStock.Text),
                     Category = txtCategory.Text,
                     VATRate = decimal.Parse(txtVATRate.Text),
